fix(api): harden chunked upload cleanup
